Created shared BrushExtensions class between iOS and macOS
