+++ Xamarin.Forms.Platform.iOS/Extensions/BrushExtensions.shared.cs
@@ -1,0 +1,67 @@
+﻿using System.Collections.Generic;
+using System.Linq;
+using CoreGraphics;
+using Foundation;
+
+#if __MOBILE__
+namespace Xamarin.Forms.Platform.iOS
+#else
+namespace Xamarin.Forms.Platform.MacOS
+#endif
+{
+	public partial class BrushExtensions
+	{
+		static CGPoint GetRadialGradientBrushEndPoint(Point startPoint, double radius)
+		{
+			double x = startPoint.X == 1 ? (startPoint.X - radius) : (startPoint.X + radius);
+
+			if (x < 0)
+				x = 0;
+
+			if (x > 1)
+				x = 1;
+
+			double y = startPoint.Y == 1 ? (startPoint.Y - radius) : (startPoint.Y + radius);
+
+			if (y < 0)
+				y = 0;
+
+			if (y > 1)
+				y = 1;
+
+			return new CGPoint(x, y);
+		}
+
+		static NSNumber[] GetCAGradientLayerLocations(List<GradientStop> gradientStops)
+		{
+			if (gradientStops == null || gradientStops.Count == 0)
+				return new NSNumber[0];
+
+			if (gradientStops.Count > 1 && gradientStops.Any(gt => gt.Offset != 0))
+				return gradientStops.Select(x => new NSNumber(x.Offset)).ToArray();
+			else
+			{
+				int itemCount = gradientStops.Count;
+				int index = 0;
+				float step = 1.0f / itemCount;
+
+				NSNumber[] locations = new NSNumber[itemCount];
+
+				foreach (var gradientStop in gradientStops)
+				{
+					float location = step * index;
+					bool setLocation = !gradientStops.Any(gt => gt.Offset > location);
+
+					if (gradientStop.Offset == 0 && setLocation)
+						locations[index] = new NSNumber(location);
+					else
+						locations[index] = new NSNumber(gradientStop.Offset);
+
+					index++;
+				}
+
+				return locations;
+			}
+		}
+	}
+}
